feat: phone number input component working correctly
