+++ package.json
@@ -24,25 +24,26 @@
         "bootstrap": "^4.6.2",
         "bootstrap-vue": "^2.23.1",
         "dot-object": "^2.1.4",
-        "luxon": "^3.4.4",
         "js-cookie": "^3.0.5",
+        "luxon": "^3.4.4",
         "nuxt": "^2.17.3",
         "pinia": "^2.1.7",
         "ts-node": "^10.9.2",
         "vue-flatpickr-component": "^8.1.8",
         "vue-infinite-loading": "^2.4.5",
         "vue-phone-number-input": "^1.12.13",
+        "vue-virtual-scroller": "^1.1.2",
         "zod": "^3.23.8"
     },
     "devDependencies": {
-        "@types/luxon": "^3.4.2",
-        "@types/lodash-es": "^4.17.12",
         "@nuxt/types": "^2.17.3",
         "@nuxt/typescript-build": "^3.0.2",
         "@nuxtjs/eslint-config-typescript": "^12.0.0",
         "@nuxtjs/eslint-module": "^3.1.0",
         "@nuxtjs/recaptcha": "^1.1.2",
         "@nuxtjs/svg": "^0.4.0",
+        "@types/lodash-es": "^4.17.12",
+        "@types/luxon": "^3.4.2",
         "@typescript-eslint/eslint-plugin": "^6.13.2",
         "@typescript-eslint/parser": "^6.13.2",
         "eslint": "^8.47.0",
@@ -55,6 +56,7 @@
         "eslint-plugin-vue": "^9.17.0",
         "husky": "^9.0.11",
         "lint-staged": "^15.2.5",
+        "lodash-es": "^4.17.21",
         "prettier": "^3.1.0",
         "sass": "1.64.0",
         "sass-loader": "^10.4.1",
@@ -63,8 +65,7 @@
         "stylelint-config-standard-scss": "^11.1.0",
         "typescript": "4.9.5",
         "vue-svg-loader": "^0.16.0",
-        "vue-template-compiler": "^2.7.14",
-        "lodash-es": "^4.17.21"
+        "vue-template-compiler": "^2.7.14"
     },
     "lint-staged": {
         "*.{js,jsx,ts,html,vue}": "eslint --fix",
